Applied inttypes.h casting consistently

+++ IRremote.h
@@ -14,6 +14,7 @@
 
 #ifndef IRremote_h
 #define IRremote_h
+#include <inttypes.h>
 
 // The following are compile-time library options.
 // If you change them, recompile the library.
@@ -26,12 +27,12 @@
 // Results returned from the decoder
 class decode_results {
 public:
-  int decode_type; // NEC, SONY, RC5, UNKNOWN
-  unsigned int panasonicAddress; // This is only used for decoding Panasonic data
-  unsigned long value; // Decoded value
-  int bits; // Number of bits in decoded value
-  volatile unsigned int *rawbuf; // Raw intervals in .5 us ticks
-  int rawlen; // Number of records in rawbuf.
+  int16_t decode_type; // NEC, SONY, RC5, UNKNOWN
+  uint16_t panasonicAddress; // This is only used for decoding Panasonic data
+  int32_t value; // Decoded value //mpf need to make unsigned.
+  int16_t bits; // Number of bits in decoded value
+  volatile uint16_t *rawbuf; // Raw intervals in .5 us ticks
+  int16_t rawlen; // Number of records in rawbuf.
 };
 
 // Values for decode_type
@@ -56,22 +57,22 @@ class IRrecv
 public:
   IRrecv(int recvpin);
   void blink13(int blinkflag);
-  int decode(decode_results *results);
+  int16_t decode(decode_results *results);
   void enableIRIn();
   void resume();
 private:
   // These are called by decode
-  int getRClevel(decode_results *results, int *offset, int *used, int t1);
-  long decodeNEC(decode_results *results);
-  long decodeSony(decode_results *results);
-  long decodeSanyo(decode_results *results);
-  long decodeMitsubishi(decode_results *results);
-  long decodeRC5(decode_results *results);
-  long decodeRC6(decode_results *results);
-  long decodePanasonic(decode_results *results);
-  long decodeJVC(decode_results *results);
-  long decodeHash(decode_results *results);
-  int compare(unsigned int oldval, unsigned int newval);
+  int16_t getRClevel(decode_results *results, int16_t *offset, int16_t *used, int16_t t1);
+  int32_t  decodeNEC(decode_results *results);
+  int32_t  decodeSony(decode_results *results);
+  int32_t  decodeSanyo(decode_results *results);
+  int32_t  decodeMitsubishi(decode_results *results);
+  int32_t  decodeRC5(decode_results *results);
+  int32_t  decodeRC6(decode_results *results);
+  int32_t  decodePanasonic(decode_results *results);
+  int32_t  decodeJVC(decode_results *results);
+  int32_t decodeHash(decode_results *results);
+  int16_t compare(uint16_t oldval, uint16_t newval);
 
 } 
 ;
@@ -87,22 +88,22 @@ class IRsend
 {
 public:
   IRsend() {}
-  void sendNEC(unsigned long data, int nbits);
-  void sendSony(unsigned long data, int nbits);
+  void sendNEC(int32_t data, int16_t nbits);
+  void sendSony(int32_t data, int16_t nbits);
   // Neither Sanyo nor Mitsubishi send is implemented yet
-  //  void sendSanyo(unsigned long data, int nbits);
-  //  void sendMitsubishi(unsigned long data, int nbits);
-  void sendRaw(unsigned int buf[], int len, int hz);
-  void sendRC5(unsigned long data, int nbits);
-  void sendRC6(unsigned long data, int nbits);
-  void sendDISH(unsigned long data, int nbits);
-  void sendSharp(unsigned long data, int nbits);
-  void sendPanasonic(unsigned int address, unsigned long data);
-  void sendJVC(unsigned long data, int nbits, int repeat); // *Note instead of sending the REPEAT constant if you want the JVC repeat signal sent, send the original code value and change the repeat argument from 0 to 1. JVC protocol repeats by skipping the header NOT by sending a separate code value like NEC does.
+  //  void sendSanyo(int32_t data, int16_t nbits);
+  //  void sendMitsubishi(int32_t data, int16_t nbits);
+  void sendRaw(uint16_t buf[], int16_t len, int16_t hz);
+  void sendRC5(int32_t data, int16_t nbits);
+  void sendRC6(int32_t data, int16_t nbits);
+  void sendDISH(int32_t data, int16_t nbits);
+  void sendSharp(int32_t data, int16_t nbits);
+  void sendPanasonic(uint16_t address, int32_t data);
+  void sendJVC(int32_t data, int16_t nbits, int16_t repeat); // *Note instead of sending the REPEAT constant if you want the JVC repeat signal sent, send the original code value and change the repeat argument from 0 to 1. JVC protocol repeats by skipping the header NOT by sending a separate code value like NEC does.
   // private:
-  void enableIROut(int khz);
-  VIRTUAL void mark(int usec);
-  VIRTUAL void space(int usec);
+  void enableIROut(int16_t khz);
+  VIRTUAL void mark(int16_t usec);
+  VIRTUAL void space(int16_t usec);
 }
 ;
